refactor(init): introduce src/init, src/list
