+++ packages/akashic-cli-init/src/index.ts
@@ -1,3 +1,3 @@
-export { promiseInit, init } from "./init";
-export { listTemplates } from "./listTemplates";
+export { promiseInit, init } from "./init/init";
+export { listTemplates } from "./list/listTemplates";
 export { run } from "./cli";
