+++ packages/akashic-cli-init/src/init/TemplateConfig.ts
