+++ packages/akashic-cli-init/src/init/copyTemplate.ts
