+++ packages/akashic-cli-init/src/init/downloadTemplate.ts
