+++ packages/akashic-cli-init/src/init/extractZipIfNeeded.ts
