+++ packages/akashic-cli-init/src/init/readTemplateFile.ts
