+++ packages/akashic-cli-init/src/init/showTemplateMessage.ts
